feat: add the exception

+++ src/main/java/com/matheusfbio/spring_practice/service/ProductService.java
@@ -1,10 +1,11 @@
 package com.matheusfbio.spring_practice.service;
 
 import java.util.List;
-import java.util.Optional;
 
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 
+import com.matheusfbio.spring_practice.excepctions.ResourceNotFound;
 import com.matheusfbio.spring_practice.model.Product;
 import com.matheusfbio.spring_practice.repository.ProductRepository;
 
@@ -19,9 +20,9 @@ public class ProductService {
     public List<Product> getAllProducts() {
         return productRepository.findAll();
     }
-
-    public Optional<Product> getProductById(Long id) {
-        return productRepository.findById(id);
+    public Product getProductById(@NonNull Long id) {
+        return productRepository.findById(id)
+        .orElseThrow(() -> new ResourceNotFound("The product with id "+id+" Não foi encontrado"));
     }
 
     public Product saveProduct(Product product) {
@@ -32,8 +33,12 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
-             if (id == null) {
+            if (id == null) {
             throw new IllegalArgumentException("Product cannot be null");
+        }
+
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFound("The product with id "+id+" Not Found");
         }
         productRepository.deleteById(id);
     }
